+++ packages/simple-jwt-login/CHANGELOG.md
@@ -20,10 +20,10 @@ This was a version bump only for simple-jwt-login to align it with other project
 
 - **simple-jwt-login:** add `revokeTokenBeforeLogin` to automatically revoke token before each new login ([079128e](https://github.com/kmakris23/store-sdk/commit/079128e))
 - **simple-jwt-login:** add `fetchCartOnLogin` to automatically fetch cart after login and automatically remove `nonce` and `cartToken` on each logout ([fb6e964](https://github.com/kmakris23/store-sdk/commit/fb6e964))
-- ⚠️  new event bus implementation ([53b0207](https://github.com/kmakris23/store-sdk/commit/53b0207))
+- ⚠️ new event bus implementation ([53b0207](https://github.com/kmakris23/store-sdk/commit/53b0207))
 - **simple-jwt-login:** automatically call `clearToken()` on token revoke success ([33c35c3](https://github.com/kmakris23/store-sdk/commit/33c35c3))
 
-### ⚠️  Breaking Changes
+### ⚠️ Breaking Changes
 
 - New event bus has been implemented replacing the previous event emitter.